Use lseek instead of lseek64 in Unix I/O
As part of this, changed interop definitions of off_t to be Int64 rather than IntPtr.  For now we're assuming that we'll run either on 64-bit systems and/or where _FILE_OFFSET_BITS == 64.  We can revisit this if that assumption fails.

+++ src/Common/src/Interop/Unix/libc/Interop.posix_fadvise.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-using off_t = System.IntPtr;
+using off_t = System.Int64; // Assuming either 64-bit machine or _FILE_OFFSET_BITS == 64
 
 internal static partial class Interop
 {
